Properly setup geant4-data package
The data files do not change with patch versions of Geant4 when using
the same minor version. Change recipe such that one always get the
same set of data files for all patch versions of a minor version.
Use the proper geant4-data versions in the geant4 recipe.
Correct the creation of the link to the data files in the Geant4 recipe.